DETR-Jittor/requirements.txt
astunparse==1.6.3
cycler==0.11.0
easydict==1.13
fonttools==4.38.0
jittor==1.3.9.14
kiwisolver==1.4.5
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.12.0.88
packaging==24.0
Pillow==9.5.0
pycocotools==2.0.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scipy==1.7.3
six==1.17.0
tqdm==4.67.1
typing_extensions==4.7.1
